Fix some compiler warnings.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/4288



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@17043 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/tools/pkgdata/pkgtypes.c
@@ -217,7 +217,7 @@ CharList *pkg_appendUniqueDirToList(CharList *l, CharList** end, const char *str
         return l; /* no dir path */
     }
     if((rPtr-strAlias) > (sizeof(aBuf)/sizeof(aBuf[0]))) {
-        fprintf(stderr, "## ERR: Path too long [%d chars]: %s\n", sizeof(aBuf), strAlias);
+        fprintf(stderr, "## ERR: Path too long [%d chars]: %s\n", (int)sizeof(aBuf), strAlias);
         return l;
     }
     strncpy(aBuf, strAlias,(rPtr-strAlias));
